requirements.txt
Django==1.7.1
Jinja==1.2
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
argparse==1.2.1
distribute==0.6.24
django-bootstrap-form==3.1
django-bootstrap3==5.0.3
django-crispy-forms==1.4.0
django-floppyforms==1.1
django-jinja==1.0.5
wsgiref==0.1.2
